+++ CPP_Module_02/ex02/Fixed.cpp
@@ -5,33 +5,30 @@ const int Fixed::fractions = 8;
 
 Fixed::Fixed(void):
 	raw(0) {
-	// std::cout << "Default constructor called" << std::endl;
 }
 
-/* Numbers above () or below () is UB */
 Fixed::Fixed(const int num):
 	raw(0) {
-	// std::cout << "Int constructor called" << std::endl;
-	this->raw = num << this->fractions;
+	this->raw = static_cast<int>(static_cast<unsigned int>(num) << this->fractions);
 }
 
 Fixed::Fixed(const float num):
 	raw(0) {
-	// std::cout << "Float constructor called" << std::endl;
+	// We scale the float num, then because casting from float to int removes the
+	//	fraction completely, we add 0.5 if it's positive so if it's closer to the
+	//	next whole number it gets truncated to that one. (3.7 + 0.5 == 3.2; gets
+	//	truncated to 3.0, then safely stored in an int).
 	this->raw = (num * (1 << this->fractions) + (num > 0 ? 0.5 : -0.5));
 }
 
 Fixed::~Fixed(void) {
-	// std::cout << "Destructor called" << std::endl;
 }
 
 Fixed::Fixed(const Fixed& obj) {
-	// std::cout << "Copy constructor called" << std::endl;
 	*this = obj;
 }
 
 Fixed&	Fixed::operator=(const Fixed& obj) {
-	// std::cout << "Copy assignment operator called" << std::endl;
 	if (&obj != this) {
 		this->raw = obj.raw;
 	}
@@ -39,12 +36,10 @@ Fixed&	Fixed::operator=(const Fixed& obj) {
 }
 
 int		Fixed::getRawBits(void) const {
-	// std::cout << "getRawBits member function called" << std::endl;
 	return this->raw;
 }
 
 void	Fixed::setRawBits(const int raw) {
-	// std::cout << "setRawBits member function called" << std::endl;
 	this->raw = raw;
 }
 
@@ -112,7 +107,7 @@ Fixed	Fixed::operator*(const Fixed& obj) const {
 	/* This multiplication will move the point 8 more bits to the left, so
 		we scale it back by shifting 8 right. */
 	long	mul = static_cast<long>(this->raw) * static_cast<long>(obj.raw);
-	ret.setRawBits(mul >> this->fractions);
+	ret.setRawBits(static_cast<int>(mul >> this->fractions));
 	return ret;
 }
 
@@ -124,7 +119,8 @@ Fixed	Fixed::operator/(const Fixed& obj) const {
 	/* We only scale the first operand to make the division more precise. Scaling both doesn't make sense,
 		because in a division operation if both sides are multiplied by the same amount, that amount is ignored
 		anyway. (100*x / 200*x == 100 / 200), but (100*x / 200) will give use more info */
-	long	div = ((static_cast<long>(this->raw)) << this->fractions) / ((static_cast<long>(obj.raw)));
+	unsigned long	div = ((static_cast<unsigned long>(this->raw)) << this->fractions)
+						/ ((static_cast<unsigned long>(obj.raw)));
 	/* the result (div) doesn't need to be shifted after this, because it's already scaled
 		(the whole and fraction parts of the bits are in their correct position) */
 	ret.setRawBits(static_cast<int>(div));
